:sparkles: 다양한 예외를 처리할 수 있도록 상위 객체 추가
